+++ app/src/main/java/ru/reliableteam/noteorganizer/notes/single_note_activity/presenter/SingleNotePresenter.java
@@ -23,7 +23,7 @@ public class SingleNotePresenter extends NoteDaoImpl implements BasePresenter {
     }
 
     public void getClickedNote() {
-        int id = appSettings.getClickedNoteId();
+        long id = appSettings.getClickedNoteId();
         if (id != NEW_NOTE) {
             getNote(id, this);
         }
